refactor: 엔티티 수정 및 패키지 구조 개선
