+++ src/main/java/balancetalk/domain/file/entity/File.java
@@ -1,13 +1,16 @@
-package balancetalk.domain;
+package balancetalk.domain.file.entity;
 
+import balancetalk.domain.Notice;
+import balancetalk.domain.file.enums.FileType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
@@ -22,7 +25,10 @@ public class File {
 
     private String name;
     private String path;
+
+    @Enumerated(value = EnumType.STRING)
     private FileType type;
+
     private Long size; // TODO 수정 필요
 
     @ManyToOne(fetch = FetchType.LAZY)
