+++ src/main/java/balancetalk/domain/file/enums/FileType.java
@@ -1,4 +1,4 @@
-package balancetalk.domain;
+package balancetalk.domain.file.enums;
 
 public enum FileType {
     JPG, JPEG, PNG
