+++ src/main/java/balancetalk/domain/member/enums/Role.java
@@ -1,4 +1,4 @@
-package balancetalk.domain;
+package balancetalk.domain.member.enums;
 
 public enum Role {
     ADMIN, USER, GUEST
