+++ src/main/java/balancetalk/domain/post/entity/Post.java
@@ -1,5 +1,10 @@
-package balancetalk.domain;
+package balancetalk.domain.post.entity;
 
+import balancetalk.domain.comment.entity.Comment;
+import balancetalk.domain.report.entity.Report;
+import balancetalk.domain.ViewStatus;
+import balancetalk.domain.member.entity.Member;
+import balancetalk.domain.post.enums.PostCategory;
 import balancetalk.global.common.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,7 +14,6 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToMany;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import java.time.LocalDateTime;
@@ -41,13 +45,13 @@ public class Post extends BaseTimeEntity {
     @JoinColumn(name = "member_id")
     private Member member;
 
-    @OneToMany(mappedBy = "balance_option")
+    @OneToMany(mappedBy = "post")
     private List<BalanceOption> options = new ArrayList<>();
 
-    @OneToMany(mappedBy = "post_like")
+    @OneToMany(mappedBy = "post")
     private List<PostLike> likes = new ArrayList<>();
 
-    @OneToMany(mappedBy = "comment")
+    @OneToMany(mappedBy = "post")
     private List<Comment> comments = new ArrayList<>();
 
     // TODO 개선 필요
@@ -55,9 +59,9 @@ public class Post extends BaseTimeEntity {
 //    @JoinColumn(name = "tag_id")
 //    private List<Tag> tags = new ArrayList<>();
 
-    @OneToMany(mappedBy = "bookmark")
+    @OneToMany(mappedBy = "post")
     private List<Bookmark> bookmarks = new ArrayList<>();
 
-    @OneToMany(mappedBy = "report")
+    @OneToMany(mappedBy = "post")
     private List<Report> reports = new ArrayList<>();
 }
