+++ src/main/java/balancetalk/domain/post/entity/PostLike.java
@@ -1,5 +1,6 @@
-package balancetalk.domain;
+package balancetalk.domain.post.entity;
 
+import balancetalk.domain.member.entity.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -12,18 +13,18 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Bookmark {
+public class PostLike {
 
     @Id
     @GeneratedValue
-    @Column(name = "bookmark_id")
+    @Column(name = "post_like_id")
     private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
 }
