+++ src/main/java/balancetalk/domain/post/enums/PostCategory.java
@@ -1,4 +1,4 @@
-package balancetalk.domain;
+package balancetalk.domain.post.enums;
 
 public enum PostCategory {
     CASUAL, DISCUSSION
