+++ src/main/java/balancetalk/domain/vote/entity/Vote.java
@@ -1,5 +1,7 @@
-package balancetalk.domain;
+package balancetalk.domain.vote.entity;
 
+import balancetalk.domain.member.entity.Member;
+import balancetalk.domain.post.entity.BalanceOption;
 import balancetalk.global.common.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
